% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_to_coursera_import.R
\name{course_to_coursera_import}
\alias{course_to_coursera_import}
\title{Convert a Coursera Course to Coursera Import}
\usage{
course_to_coursera_import(x)
}
\arguments{
\item{x}{object of class \code{coursera_import} or
\code{coursera_course_import}}
}
\value{
object of class \code{coursera_import}
}
\description{
Convert a Coursera Course to Coursera Import
}
