% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_import_as_course.R
\name{crsra_import_as_course}
\alias{crsra_import_as_course}
\title{Convert a Coursera Course to Coursera Import}
\usage{
crsra_import_as_course(x)
}
\arguments{
\item{x}{object of class \code{coursera_import} or
\code{coursera_course_import}}
}
\value{
object of class \code{coursera_import}
}
\description{
Convert a Coursera Course to Coursera Import
}
