% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_tabledesc.R
\name{crsra_tabledesc}
\alias{crsra_tabledesc}
\title{Description for a table}
\usage{
crsra_tabledesc(x)
}
\arguments{
\item{x}{The name of the table}
}
\value{
The description for a table based on the description provided by Coursera in the data exports
}
\description{
Description for a table
}
\examples{
crsra_tabledesc("assessments")
}
