% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_timetofinish.R
\name{crsra_timetofinish}
\alias{crsra_timetofinish}
\title{Time that took each learner (in days) to finish a course}
\usage{
crsra_timetofinish(all_tables)
}
\arguments{
\item{all_tables}{A list from \code{\link{crsra_import_course}} or
\code{\link{crsra_import}}}
}
\value{
A table containing \code{hashed_user_id}s with a column indicating the time (in days) that took each user to complete a course. The time is calculated as the difference between the last and first activity in the a course.
}
\description{
Time that took each learner (in days) to finish a course
}
\examples{
crsra_timetofinish(example_course_import)
}
