% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_course_import.R
\docType{data}
\name{example_course_import}
\alias{example_course_import}
\title{Example Import of a Coursera Course}
\format{A list with 100 elements, which are \code{data.frame}s
imported from a fake Coursera class:}
\usage{
example_course_import
}
\description{
Example Import of a Coursera Course
}
\keyword{datasets}
