% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_course_import.R
\docType{data}
\name{tabdesc}
\alias{tabdesc}
\title{Table Descriptions}
\format{A vector table descriptions, where the names
of the table descriptions is the name of the
tables in an import.}
\usage{
tabdesc
}
\description{
Table Descriptions
}
\keyword{datasets}
