% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_assessmentskips.R
\name{crsra_assessmentskips}
\alias{crsra_assessmentskips}
\title{Frequencies of skipping an peer-assessed submission}
\usage{
crsra_assessmentskips(bygender = FALSE, wordcount = TRUE, n = 20)
}
\arguments{
\item{bygender}{A logical value indicating whether results should be broken down by gender}

\item{wordcount}{A logical value indicating whether word count should be shown in the results; default is true}

\item{n}{An integer indicating the number of rows for the word count}
}
\value{
The outputs are frequency tables (tibble).and are shown for each specific course
}
\description{
Frequencies of skipping an peer-assessed submission
}
\examples{
crsra_assessmentskips()
crsra_assessmentskips(bygender = TRUE, n = 10)
}
