% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_import.R
\name{crsra_import}
\alias{crsra_import}
\title{dbListTables(con)
dbListFields(con, 'peer_submissions')}
\usage{
crsra_import(rmd = TRUE, workdir = getwd())
}
\description{
dbListTables(con)
dbListFields(con, 'peer_submissions')
}
