% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_import.R
\name{crsra_import}
\alias{crsra_import}
\title{Imports all the .csv files into one list consisting of all the courses and all the tables within each course.}
\usage{
crsra_import(rmd = TRUE, workdir = getwd())
}
\arguments{
\item{workdir}{A character string vector indicating the directory where all the unzipped course directories are stored.}
}
\description{
Imports all the .csv files into one list consisting of all the courses and all the tables within each course.
}
\examples{
crsra_import()
}
