% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_membershares.R
\name{crsra_membershares}
\alias{crsra_membershares}
\title{ggplot(temp, aes(course_membership_role)) + geom_bar(aes(weight = freq))}
\usage{
crsra_membershares(groupby = "roles")
}
\description{
ggplot(temp, aes(course_membership_role)) + geom_bar(aes(weight = freq))
}
