% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsra_tabledesc.R
\name{crsra_tabledesc}
\alias{crsra_tabledesc}
\title{Returns description for a table}
\usage{
crsra_tabledesc(x)
}
\arguments{
\item{x}{Name of the table to get the description}
}
\value{
The description for a table based on the
description provided by Coursera in the data exports
}
\description{
Returns description for a table
}
\examples{
crsra_tabledesc("assessments")
}
